+++ Source/Scene/GltfImageLoader.js
@@ -276,7 +276,9 @@ function loadImageFromUri(resource) {
     return loadCRN(resource);
   }
   // Resolves to an ImageBitmap or Image
-  return resource.fetchImage();
+  return resource.fetchImage({
+    skipColorSpaceConversion: true,
+  });
 }
 
 /**